move common test function to utils file.
